requirements.txt
altgraph==0.17.4
importlib_metadata==8.7.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
zipp==3.23.0
